Split airport.rs (598 lines) into airport/ subdirectory
Split crates/simulation/src/airport.rs into focused modules:
- airport/mod.rs: thin facade with mod declarations and re-exports
- airport/tier.rs (68 lines): AirportTier enum and methods
- airport/stats.rs (21 lines): AirportStats resource
- airport/systems.rs (144 lines): update_airports system and AirportPlugin
- airport/tests.rs (351 lines): all unit tests

All public items re-exported from mod.rs — no external import changes needed.
